[BE] refactor(#268): 피드백반영
* 이벤트 두개가진 리스너 행동분리
