+++ backend/src/main/java/com/example/backend/domain/define/fcm/listener/ApplyMemberListener.java
@@ -2,7 +2,6 @@ package com.example.backend.domain.define.fcm.listener;
 
 
 import com.example.backend.domain.define.fcm.FcmToken;
-import com.example.backend.domain.define.study.info.event.ApplyApproveRefuseMemberEvent;
 import com.example.backend.domain.define.study.info.event.ApplyMemberEvent;
 import com.example.backend.study.api.event.FcmSingleTokenRequest;
 import com.example.backend.study.api.event.service.FcmService;
@@ -16,7 +15,7 @@ import org.springframework.stereotype.Component;
 @Component
 @RequiredArgsConstructor
 @Slf4j
-public class StudyEventListener {
+public class ApplyMemberListener {
 
     private final FcmService fcmService;
 
@@ -32,32 +31,6 @@ public class StudyEventListener {
                 .token(fcmToken.getFcmToken())
                 .title("[" + event.getStudyTopic() + "] 스터디 신청")
                 .message(event.getName() + "님이 스터디를 신청했습니다.\n" + "프로필과 메시지를 확인 후, 수락해주세요!")
-                .build());
-    }
-
-
-    @Async
-    @EventListener
-    public void applyApproveRefuseMemberListener(ApplyApproveRefuseMemberEvent event) throws FirebaseMessagingException {
-
-        FcmToken fcmToken = fcmService.findFcmTokenByIdOrThrowException(event.getApplyUserId());
-
-        String title;
-        String message;
-
-        if (event.isApprove()) {
-            title = "[" + event.getStudyTopic() + "] 스터디 신청";
-            message = String.format("축하합니다! '%s'님 가입이 승인되었습니다!", event.getName());
-
-        } else {
-            title = "[" + event.getStudyTopic() + "] 스터디 신청";
-            message = String.format("안타깝게도 '%s'님은 가입이 거절되었습니다.", event.getName());
-        }
-
-        fcmService.sendMessageSingleDevice(FcmSingleTokenRequest.builder()
-                .token(fcmToken.getFcmToken())
-                .title(title)
-                .message(message)
                 .build());
     }
 }
